Amira: fix reading of > 4 GB files
Fixes ticket #10623.

+++ components/bio-formats/src/loci/formats/in/AmiraReader.java
@@ -105,7 +105,7 @@ public class AmiraReader extends FormatReader {
       }
     }
     else {
-      in.seek(offsetOfFirstStream + no * planeSize);
+      in.seek(offsetOfFirstStream + (long) no * planeSize);
       readPlane(in, x, y, w, h, buf);
     }
